requirements.txt
Pillow
djangorestframework
requests
django_cron
django-cors-headers
django-social-share


